requirements.txt
pandas==1.2.4
google-cloud==0.34.0
google-cloud-vision==2.3.1